The travelling purple cube (Vec3 is now Double)

+++ common/src/main/kotlin/elements/SerializableNode.kt
@@ -7,5 +7,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class SerializableNode(val id: Long, val tags: List<SerializableTag>, val coords: Vec3) {
-    constructor(wrappedNode: OsmNode) : this(wrappedNode.id, wrappedNode.tags.map { SerializableTag(it) }, Vec3(wrappedNode.latitude.toFloat(),0f, wrappedNode.longitude.toFloat()))
+    constructor(wrappedNode: OsmNode) : this(wrappedNode.id, wrappedNode.tags.map { SerializableTag(it) }, Vec3(wrappedNode.latitude,0.0, wrappedNode.longitude))
 }
